Flight connect for Employee

+++ src/main/java/Controller/FlightManagerController.java
@@ -24,9 +24,11 @@ public class FlightManagerController {
     public void deleteBtn() {
         flightManager.getDelete().setOnAction(actionEvent -> {
             Flight flight = flightManager.getTable().getSelectionModel().getSelectedItem();
-            Database.delete(flight);
-            FlightManagerController flightManagerController = new FlightManagerController();
-            this.flightManager.getScene().setRoot(flightManagerController.getFlightManager());
+            if (flight!=null) {
+                Database.delete(flight);
+                FlightManagerController flightManagerController = new FlightManagerController();
+                this.flightManager.getScene().setRoot(flightManagerController.getFlightManager());
+            }
         });
     }
